+++ src/Poc.Order.Api.Infrastructure.Data/Repositories/PedidoRepository.cs
@@ -42,7 +42,7 @@ namespace Poc.Order.Api.Infrastructure.Data.Repositories
             return mapper.Map<List<Pedido>>(listPedidoModel);
         }
 
-        public async Task UpdateImpostoPedidoAsync(int id, double imposto, CancellationToken cancellationToken)
+        public async Task UpdateImpostoPedidoAsync(int id, decimal imposto, CancellationToken cancellationToken)
         {
             var filtro = Builders<PedidoModel>.Filter.Eq(p => p.PedidoId, id);
             var update = Builders<PedidoModel>.Update
